+++ athetes_unlimited_py/__init__.py
@@ -1,10 +1,10 @@
-############################################################################## 
-##
-##  Imports
-##
-##  In the future, this will be explicitly imported.
-##
-############################################################################## 
+##############################################################################
+#
+#  Imports
+#
+#  In the future, this will be explicitly imported.
+#
+##############################################################################
 
 from athetes_unlimited_py.basketball import *
 from athetes_unlimited_py.lacrosse import *